First commit - add modules
This commit adds the modules 8-11 to the project and adds the content
folder as well.

The modules include the os, path, and file sync (fs) built in modules.

The content folder contains some text files used to showcase some
of the use cases for the modules such as reading/writing files.
Concatenation of files. And the difference between sync and async
functions.