Add middleware
This commit adds the middleware section of the tutorial.

The authorize.js module is an example of importing an authorization
module.
The logger.js module is an example of our ability to make our own
custom logger functions. In this case we take in the request method,
requested url, and the year and log it to the console.
Morgan is an example of bringing in a third party logger to use in our
app.

This demonstrates how middleware can be used to interpret request
data and be used in logic and authorization for response