Add controllers and routes
This commit adds the controllers and routes folders.
This breaks the project up into a more logical structure to maintain
a friendlier code base.

Furthermore our code can be broken up and refactored to make our
code more readable.

With the addition of the javascript.html file we can accept login
requests.  To parse this information we use urlencode and json modules
provided by express.

By adding controllers we are able to interpret user interaction
on the webpage and informs the model and view to change when needed.

We showcased the capabilities by adding a user to a list and presenting
the updated list back to the user.